Use the new incoming call confirmation APIs.
Update Telecomm and Telecomm test package to use the new
confirm-incoming-call APIs including confirmIncomingCall and
handleConfirmedIncomingCall.  Test call service always confirms all
incoming calls.

Also adds a list of unconfirmed incoming calls to the adapter. The
purpose of this is for the adapter to know which incoming call
confirmations to accept and which to ignore.

Change-Id: I914f25c45e5c887ca84d181d3dc173f25577e520