Handle disconnect during handoff
Currently the handoff code expects that the original call will stay
alive during the handoff. If the original call disconnects then
handoff is cancelled and both calls are removed.

With this CL, if the original call disconnects during handoff it's
kept alive. If the handoff fails then the both calls are removed.
If handoff succeeds then the original call is updated with the
handoff call's state.

Change-Id: Ie4a6d38d990812cc1e83c81502c008bcf9a35053