Temp fix for handoff on CDMA
With this CL we no longer disconnect the WiFi call when doing
a handoff from WiFi to to CDMA. Since CDMA becomes active
quickly, disconnecting the WiFi call will end the call and stop
handoff.

This is a temp fix. I'll implement a proper fix when we switch
to daisy chaining.

Bug: 15019557
Change-Id: I6e5d21bafd41948c8bd429740e36f31ab81deadc