Fix #911 calls.
We were previously calling Uri.parse() on the number which caused the #
to be recognized as a uri fragment separator when it should have been
part of the scheme-specific-part. Later on when we got the scheme
specific part to actually dial the number, we got an empty string.

The fix is to use Uri.fromParts() to build the string with the number
being explicitly part of the scheme-specific part.

Bug: 15315198
Change-Id: I1231d446c6ba98df417a83843181957a4d788918