Remove timeout for outgoing calls.
We added a timeout for outgoing calls previously to protect from a
connection service attempting a call forever.  However, we have since
implemented an "abort()" method which allows a user to stop an outgoing
call at any moment with the END-CALL button. This removes the need
timeout the calls since the user will always have immediate control.

It also allows apps to perform other operations that may be required as
part of the call() method that may take more than the previously
specified 60-second timeout. A selection UI to pick an account, for
example.

Bug: 15885208
Change-Id: I1450d1e29b64cc6b128c2b017e2cfa0c3a47106a