Change isRinging and isInAPhoneCall to execute from binder thread.
We call into the audio service from telecomm main thread through
setMode, but it calls us back on isInAPhoneCall which tries to move
the code execution to the main thread (where it waiting for
audio service) causing a deadlock.  Change isInAPhoneCall() to support
multi-threaded execution and do not force the invocation onto the
main thread.

Bug: 16179606
Change-Id: I2cc6f6220b34a2b35841a08897a50f2dc50839c6