Do not use stale PhoneAccounts when placing calls.
This is important in case the call log has a record of a phone account
which is no longer valid.  An example would be the emergency
PhoneAccount. That will always fail for non-emergency numbers...but if
it is the only account, then it will be used and fail. When the user
pops in a valid SIM, it will continue to fail because the phone account
saved into the call log was the stale emergency phoneAccount.

Without this fix, the user can get "Call not sent" frequently.

Bug: 16526102
Change-Id: Ic08768b85d374e6f7ab54f34b5caf501b061d49b