VT preview camera setup. (1/2)
Changing the means by which the preview camera is set for video calls.
Previously, an assumption was made that as soon as the camera is set
on the provider, the camera capabilities would be returned.
Changed so that instead the InCall-UI will explicitly request the
camera capabilities after the camera is set to match the original
intended camera setup flow.

Change-Id: I539e6bd72404e06f89f2e684cbb40c51637ead19