Changing the point where the call state is set.
The reason speakerphone was not working while the call was dialing was
because the call went from state DIALING -> DIALING, which was not seen
as a state change, which meant that the call was not updated in the audio
manager, resulting there being no audio focus.

Note this is related to changes made in CL/518284. 

Bug: 16978516
Change-Id: I7b2be0ec21c484f6fbd767055e5e9afefbb526e2