Fix PhoneAccountPreferences not being saved
* Add a dummy PhoneAccountHandle, which is used to differentiate
between the user selecting a null PhoneAccountHandle, and no
preference being present at all.

* Refactor getDefaultOutgoingPhoneAccount slightly and add
getUserSelectedOutgoingPhoneAccount. getUserSelectedOutgoingPhoneAccount
is only used by PhoneAccountPreferencesActivity to determine the user's
true selection. getDefaultOutgoingPhoneAccount will continue to behave
the same, and is the method that will be continue to be used to determine
the correct default PhoneAccount to use.

Bug: 17140770

Change-Id: I5f5b4940d3babde05d563024ee9618d5d65204ea