Pass voicemail: URIs directly to Telephony. Rename handle -> number
voicemail calls should be transformed to tel: numbers through Telephony.
TelephonyConnectionManager is able to pick the right number because it
has knowledge of the subscription being used for the call.

Also rename "handle" to "number" in all cases where "handle" was
referring to the phone number within a URI. The code referred to
URI "handles" as well as phone number "handles", which made it confusing
to differentiate.

Additionally removed an incorrect comment and parsed CallStates in log
messages for easier readability.

Bug: 17317190
Change-Id: I8269adef76d7770f8ef180f97c75d72b31b0bdd8