Forward CallActivity invocations to the primary user.
Telecom framework runs as the primary user only.

When CALL intents are sent via startActivity, they invoke
telecom.CallActivity on the current user (activities don't allow
specifying them as singleUser="true"). This code checks to see if
CallActivity was invoked from as the secondary user and if so
forwards the intent to the primary user's version of CallActivity.
CallActivity then simply calls directly into the rest of telecom.

NOTE: Telecom already does the right thing when binding to the
In-Call UI so that the UI always shows up under the current user.

Bug: 17253031
Change-Id: Iea7ecf4aa48ed80710a1ed35603cbf4588581afc