Refuse outgoing call if max outgoing exceeded.
This bug was happening because multiple calls in the CONNECTING state
were being added. Now, if there's an attempt to place multiple calls,
ignore when the maximum number of outgoing calls is exceeded.

Exception: if the outgoing call is an emergency call, cancel the
current outgoing call if the current outgoing call is not an
emergency call itself.

Bug: 17610917
Change-Id: I0ae041fa3a09a2fb9508f2342894e77b87001b43