Split CallActivity into trampoline activity and CallReceiver
Move most of CallActivity logic into CallReceiver. Only significant logic
remaining in CallActivity is checking for OUTGOING_CALL_RESTRICTED and
showing a toast.

Also fixes a bug where the secondary user would never be able to make
a call if the primary user had Google Voice installed, because the
ordered broadcast is sent out for the primary user, allowing Google
voice to cancel it.

Bug: 17579886

Change-Id: Id1689f36503a74e5754b10d019a8b64024d27f77