Don't create conferences with invalid calls
When a remote conference was created
ConnectionServiceWrapper would actually create 2
conferences. One for the remote connection service with no
valid calls one for the real connection service with valid
calls.

When the conference was destroyed the remove call on the
remote connection serivce would trigger a unbind causing
android.telecom.ConnectionService to end call connections
for that service.

Fix was to ensure that we only create conferences with
valid calls.

Bug: 17632595
Change-Id: I3f339120dde8eb07ee7e40a26d3a37f1dbd1409e