Turn on Bluetooth automatically for outgoing calls.
For outgoing calls, we introduced a state before DIALING called
CONNECTING.  This indicated that we were attempting to place the call
but that there wasn't an indication that it was dialing quite yet.

The CallAudioManager code which sets the initial state for new calls
was still looking for DIALING as the initial state before opting
to turn on blueooth on behalf of the user. Add CONNECTING as a state
which does that, too.

Bug:17661926
Change-Id: I7c27d0a462b7933cbf0276332b1d9f63d100f345