Ensure bluetooth does not get hangup sound when calling through gvoice
Because Google Voice works by canceling an outgoing call and starting up
its own outgoing call, telecom registers the first call as a
disconnecting call and sends a signal to the bluetooth device that the
call was disconnected. This causes a hangup sound from the bluetooth
device. This fix works by having bluetooth ignore a call until it gets
into the dialing state, which only happens after the google voice call
has connected.

Bug: 17662993
Change-Id: I52c504f615534c59ede8c0de520f755484922730