Don't stop outgoing calls if we are the foreground call.
When attempting outgoing calls we do checks to make sure there is room
for another outgoing call (we only support 1 at a time).  But that code
needs to check to see if the new call is already the foreground call and
if so, let it continue.

Without this change, we never make an outgoing call in the "Ask
First" case

Bug: 17672699
Change-Id: I9bf37b08e54e2fc2ace1295ce269775cb14d08f1