Removed failed outgoing calls.
The user is unable to add a call in this bug, because the call is not
add capable. It fails the isAddCallCapable check because there are
multiple calls; one of which is not the same call and does not have
a parent.

This appears to be because of some interaction with Google Voice (at
least in the specific case I was observing), where the initial call
is disconnected and then the call is rerouted. When this initial call
is disconnected, it was not removed, leading to this (erroneous) two
call state.

Fixing this by explicitly calling remove call for failed outgoing
calls. Formerly, they were only marked as disconnected.

Bug: 17731082
Change-Id: I0d6401cbc593a128483072c5faa189949e73d71e