Dont clear serviceConnections on subsequent bind() calls.
Bind() gets invoked with each call that is added before the first bind()
is complete. We were clearing the service connection set each time which
effectively meant that we were dropping the previous bind requests on
the floor.  The rest of the code already does the right checks to not
rebind to the same service so just remove the clear().

Bug:17735448
Change-Id: I614225b05fa29b143604fe4dfedaa1e417dc2d2d