Enforce CALL_PRIVILEGED check at start of CallActivity
This prevents third party apps from using the CALL_PRIVILEGED action with
an explicit intent to place emergency phone calls for secondary users.

The original component check did not work for secondary users because the
component name was being clobbered and replaced with the CallReceiver's
component name.

Also make sure to start the system dialer as the UserHandle.CURRENT instead of
UserHandle.OWNER so that it shows up if the user is not the primary user.

Bug: 17920957
Change-Id: Ie2aba734205361bf780f1c81ae2709da4162d4f5