Issue: turning on bluetooth headset in call doesn't change route (1/2).
Forcing queryPhoneState to send CallState to connected bluetooth headset.
The method updateHeadsetWithCallState, which queryPhoneState calls, checks
to see if any state information has changed, and then updates the headset
with the new state information.  In HeadsetStateMachine, there are calls
to queryPhoneState made when the headset is connected.  In response to the
change in call state, the headset will automatically connect or not.
However, since updateHeadsetWithCallState was not detecting a change in
the call state, the update was never sent to HeadsetStateMachine, which
meant that it would never attempt to turn on the audio for the headset.

Bug: 17933664
Change-Id: I9270a20e2d227d90f5401e0f237a908b192fc174