Don't play call waiting for first incoming call
The first incoming call causes the ringer to play call waiting
for a split second before the ringtone sounds because it is
not in the foreground at that point in time.

This causes issues with the audio layer as going to IN_CALL right before
MODE_RINGTONE results in a loss of outgoing audio for the first call.

Add a special check to avoid this scenario.

Bug: 17961223
Change-Id: I580f9c8fa21e375d0c1d6fada80aa23d1ecdb913