Fixing issue where Google Voice calls are logged twice to call log.
1. Changed disconnect cause for aborted calls to "CANCELED" which ensures
they are not logged.
2. Also, added a check in Telecom to avoid bringing up the InCall UI
immediately when a new call is made, if Google Voice is installed.  This
is to ensure that the user does not see two reveal animations.  This was
due to the fact that when the call was initially added and subsequently
aborted, the InCall ui would be shown.   With this change, if GVoice is
installed, adding the call is delayed.

Bug: 17961478
Change-Id: Ic0055a53ac85980fa22136a5c8b6cb59be347da9