Once responded to a call, remove it from list of ringing calls
Remove a call from mRingingCalls immediately once it is responded to.
This fixes an erroneous transition as described below
1) Call waiting is received
2) Call waiting is answered
3) Currently active call is put on hold
4) onForegroundCallChanged is called with the new call waiting, before it
goes to active
5) Ringtone is played in Ringer.startRingingOrCallWaiting
This CL ensures that if 4) happens, we do not erroneously transition to
AudioManager.MODE_RINGTONE.

Tested with the following scenarios to ensure that the ringtone/call waiting
tone is correctly sounded, and invalid audio mode transitions are not
encountered.

1) Multiple MT calls in a row
2) MT call + call waiting + reject
3) MT call + call waiting + answer
4) Conference call + call waiting + reject
5) Conference call + call waiting + answer

Bug: 18058600
Change-Id: I97cd798656cc56c937964fb1ebb3032c4cee4ed7