Remove ADD_CALL if not only top-level call.
ADD_CALL now depends on the state of the underlying connections, and
the mix of GSM/CDMA/IMS connections.

I propose to address this by setting those capabilities as true on
the GSM/CDMA connections, then further filtering in Telecom depending
on the call state (ie. whether it is the only top level call, and
whether or not we're dialing). I think this logic should be simpler,
and also allows us to more directly control this capability depending
on the connection type.

Bug: 18062269
Change-Id: Ie5be502545a580719e65595e495ff30cbc3fc22f