Reuse calls canceled through NEW_OUTGOING_CALL broadcast.
When a call is canceled because of NEW_OUTGOING_CALL broadcast,
we no longer remove it immediately.  This code adds a 200ms delay.
The 200ms delay is long enough that if an app reposts another call
to the same number (with a gateway number, for example), instead of
destroying and recreating the call, we attempt to reuse the existing
one. This eliminates the jank in the UI caused by the call being
destroyed and a new one recreated.

Bug: 17692408
Change-Id: I225a4fcc6f00d764cf76595f15f6d2d0785b9fe2