Fixing issue where VoLTE conference calls have incorrect call time.
1. In constructor for "Call", setting connect time for new conferences;
the connect time is normally set when a call goes ACTIVE, but since a
conference starts active it never gets a state change to active.
2. Changed code in InCallController which finds oldest connected child
call to handle the potential case where no children have a connect time
defined.  This can occur when an IMS conference has just started and
conference event package participants are added.  Since they are initially
added as "NEW", it is possible for there to be no active child call, which
could cause the timer to glitch until the child calls become active.

Bug: 18511011
Change-Id: I36782db8f89299c31e6f489d34a83701855d30c8