Fixing unhold in CallsManager so it does not hold child calls.
There was an issue (see bug) where a conference could not be held in order
to swap with a held single party call.
Looking at how the CallsManager code works, we hold all calls including
the child calls of a conference.  In Telephony, the TelephonyConference
code invokes hold on the first child-connection in the conference.
Changed the code so that the child calls are not held in Telecom.

Bug: 18793259
Change-Id: I83cde7be1e006fbd05ba79c13ee1a1138b1ae593