Unplugging headset shouldn't stop Bluetooth.
When the headset was unplugged, we changes audio to earpiece even if the
audio was currently going through Bluetooth.

This change makes it so that we keep the current routing scheme when a
headset is unplugged unless the audio was actually going through the
wired earpiece.  In that case, we go to either earpiece or speaker
depending on whether speaker was on prior to going to the wired headset.

Bug: 18823515
Change-Id: Ib12d669c2bcd8f4c8a95d450af6bcc28f15122f3