Remove line to allow hiding SMS response from InCallUI in multiSIM.
We sometimes will want to hide the SMS response option from the
InCallUI even if the subscription is able to send SMS messages. For
example in a multi-SIM phone where a call comes in to the non-default
SMS phone account. Remove a somewhat duplicate case of adding
CAPABILITY_RESPOND_VIA_SMS to handle this case.

Bug: 19034404
Change-Id: Ia2793c26a362de317e09ea0c5ea07f14fe326794