Don't log the Emergency Phone Account in the call log.
The emergency phone account is there because telecom system requires a
phone account to make calls (even emergency calls).  It gets registered
when there is no SIM and unregistered when the user adds a SIM.  It's a
place holder for the user to know that only emergency calls can be made
while there is no SIM.  A side effect is that when the user attempts to
place a regular call when there is no SIM, we end up logging "Emergency
Calls" along side the dial attempt in the call log.  This gives a false
impression that the call was an emergency call. This change makes it so
that the emergency phone account is never logged as part of the calllog.

Bug: 19569750
Change-Id: I842edd5d781dc8f4f1cc8fc6a12d2350063a1f82