Cleanup of Video Call pause functionality.
Whether the paused videoState is available is dependent on the carrier's
implementation of the VT spec. The original VT implementation assumed
that this was stored in a system property accessed via InCall; these CLs
move this to a Call/Connection capability which will ultimately support
multisim video capable devices.

- Added mapping between Connection/Call capability.
- Changed toString for VideoState in a Call so that it is easier to see
the video state of a call in the logs.

Bug: 16680364
Bug: 19820114
Change-Id: I3840f92270100811161120dffcfe297bef7c4ea2