Support for multiple VideoCall.Listeners for a VideoCall.
The current code assumes that only a single instance of VideoCall will be
provided to the default system InCall UI.  Ideally multiple
InCallService implementations should be able to use the VideoCall APIs.
Note: it only really makes sense for a single InCallService to get/set
the video surfaces.

- Modify InCallController to only update the VideoProvider in the parcelled
call data if it has changed.
- In InCallController, removed check that would prevent sending video
provider to other InCallServices other than the default incall UI.
- Created TestCallList and modified TestInCallServiceImpl to track all
calls and output some debug information about the calls/video callbacks.
- Added a new Test intent, ACTION_SEND_UPGRADE_TO_VIDEO_REQUEST which will
cause the TestInCallServiceImpl to loop through all telecom calls and
trigger the VideoCall API to send an upgrade to video request.  THis is
to mimic how the telephony test team will be using a test InCallService
to control the device.

Bug: 20092420
Change-Id: Ib781551d5168bc095f58b7b62906a9e59e43958e