Fixing issue where some video calls are not logged as video calls.
Some video calls are not properly logged as video calls in the call log.
This is due to the fact we only track video state history for the call
log when the call is active.
1. Incoming video call answered as video -- since the video state never
changed while the call is active, the fact its a video call is missed.
2. Rejected and missed calls -- calls never were active so are not tracked
as such.

Bug: 20689156
Change-Id: Ie1c155db968844a346493c78f63c055dde6b5195