Add a slew of clearCallingIdentity() invocations to telecom.
We used to use Handlers to move things onto the telecom main thread when
they came in through Binder calls.  This also had the effect of
switching to telecom's calling identity.

Since we no longer do that, we need to manually reset the calling
identity to avoid the permissions failures that we've been seeing
lately.  This change adds all the checks that are necessary today and
checks which would help down the road if any code changes.

Bug: 20830275
Change-Id: I73e3e2d310a505ac96abe3aed7f2118346c7be1f