Remove vt emergency call barring code from Telecom.
Removing code in Telecom to prevent dialing emergency calls as a video
call (this is moved to frameworks/opt/telephony).
Also ensured video emergency calls are not blocked when TTY is enabled;
filing follow-up bug 21474288 to ensure this gets reviewed again as I
think we want to ensure this is the correct behavior.

Bug: 20089489
Change-Id: I6a5f54cd6280f67b273dc0939c35fab353ad62b0