Play tone when video upgrade request is received.
Added a VideoProviderProxy which intermediates between the VideoProvider
in the ConnectionService and the InCallService.

Changed CallsManager to listen for session modification requests via the
VideoProviderProxy and report these to its listeners via the
CallsManagerListeberBase.

Changed InCallToneMonitor to listen for session modification requests and
to play a tone when a request to upgrade is received.

Also, added an intent in the TestInCallService which I neglected to commit
in a previous CL.

Bug: 20232310
Change-Id: I1b105968f519ff6b166bbd02cb57b1cf68d2cea9