Telecom: sync up behavior with Bluetooth service
When there's an incoming call, HeadsetStateMachine will actually suspend
a2dp media streams. In regular ringer mdoe, this works fine, as the
media will always duck(or pause itself) because there is an audio request change.

When the device is on priority/silent mode this never happens, so the
media will keep playing a potentially suspended a2dp stream, getting out
of sync, potentially causing the track to jump forward.

Change-Id: I3ddd51b843ab880c614778fbed6dfdd8edb68520
Signed-off-by: Roman Birg <roman@cyngn.com>