Correct initial audio state logic in CallAudioManager.
- Moved logic to trigger setInitialCallState from onCallUpdate to
updateAudioStreamAndMode.  In the case of IMS calls, the transition in
audio stream was happening in the call to updateAudioSteamAndMode
triggered from setIsRinging.  This meant we'd never call setInitialState.
Since updateAudioStreamAndMode is ultimately where the change is
happening, I figured it made sense to move the logic there.
- Added toString methods for stream and mode descriptions so I could more
easily discern what was happening in the logs.

Bug: 21555905
Change-Id: I0bccf0e7881abfb2f1a195fc0f87b3764107ed16