Fix for startActivity(ACTION_CALL)
Apps that use startActivity(ACTION_CALL) will currently always
fail to have the CALL_PHONE permission when checked in Telecom
because the package name is always null at this point.

We already check for the permission before this so if the app
does not have the permission it will already be rejected with
a SecurityException.

Bug: 22071972
Change-Id: Ibebadd052a13652d8390c8a30f18441110485336