Pass current call into CallAudioManager#setIsRinging.
It is possible that when setIsRinging() is called on CallAudioManager
that the initial audio state will be set at that point.  However,
setInitialAudioState requires the current call to determine the appropriate
audio route.  It defaults to wired or earpiece if there is no call.

Fixed by ensuring setIsRinging takes a call as context so that the initial
audio state can be properly set.

Bug: 22515446
Change-Id: I9c01cf6cf27e916d7e4117aa13e603682a5059fb