Migrate existing SIP PhoneAccount IDs to username.
We've changed the SIP phone accounts to use the SIP username as
account handle ID in M (ag/685031). So, need to migrate all existing
SIP phone accounts when upgrading from L or lower releases to use the
new account handle ID.
Adding a new version of the phone-account-registrar-state.xml file to
detect the need to migrate and replace the SIP URI's on the accounts
with only the username.

BUG: 23028921
Change-Id: I8cc6f099ec9cafb71c5cf6d040245ddb6a3b00ea