Reject incoming calls when dialing a outgoing call
When an outgoing call is still being dialled, we cannot hold that call
if a new incoming call is received at the same moment. Hence reject any
incoming calls if we're still dialing an outgoing call. These incoming
calls will be logged as missed to the user in call logger.

BUG: 23428950
Change-Id: I97693afd443073ea9642d69aa67542ee1aa3eff7