Update CallsManager state before invoking callback
We should update CallsManager internal state before informing the Calls
Manager event listeners for any call addition/removal/state-change. This
will ensure that we have the correct state when one of the listeners
query CallsManager after being informed about an event.
This will ensure for example that the background call is brought to
foreground when a call is disconnected and so call audio manager will
not relinquish the audio focus when an IMS conference is created
(b/23759265).

BUG: 23786790
Change-Id: Iccb74ac61166f1218f36b7b98df5d35a015e0c43