Fix TelecomSystemTest#testAudioManagerOperations
testAudioManagerOperations was setting the microphone to mute
and attempting to verify that the setMicrophoneMute methond on
the AudioManager was invoked. However, CallAudioManager bypasses
AudioManager and calls AudioService directly.

This change modifies CallAudioManager so that a mock AudioService
can be injected. The test was modified so that it checks that the
appropriate mute method was called on the mock AudioService instead
of checking the mock AudioManager.

Bug: 25158760
Change-Id: I0d5946ae6b4cd31c9067e818f20dcc37a2e15c86